Ensure that we can edit a resource
Previously, if we tried to edit a resource, we got an error message
saying that the File object doesn't respond to delete, since we are
deleting the previous image when we update it.

This commit implement the File#delete method which is a wrapper around
the DropboxClient#file_delete method.